Improve Cell struct
Move the struct to a separate file (with its tests).
Use the new Position struct instead of old y, x attributes.
Update main to support the new structs.
