+++ src/printer.rs
@@ -1,4 +1,4 @@
-use Cell;
+use cell::Cell;
 
 pub struct CliPrinter;
 
@@ -15,5 +15,22 @@ impl CliPrinter {
             color = White;
         }
         println!("{}", Style::new().on(color).paint("  "));
+    }
+}
+
+#[cfg(test)]
+mod cli_printer_tests {
+    #[test]
+    fn printer_print_cell() {
+        use printer::CliPrinter;
+        use cell::Cell;
+        use position::Position;
+
+        let printer = CliPrinter;
+        let position = Position::create_2d(1, 1);
+        let cell = Cell{position: position, alive: true};
+        printer.print_cell(cell);
+        let cell = Cell{position: position, alive: false};
+        printer.print_cell(cell);
     }
 }
